requirements.txt
paho-mqtt
SpeechRecognition
pyttsx3
openai
python-dotenv
pyaudio